CDD-1458: Minor code cleanup:  - removed un-necessary parenthesis around action builders in controller  - split out action builders into there own files in new `actionsbuilders` package
